--- src/timetracker/TimesheetMain.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import './TimesheetMain.css';
-import MainContent from './MainContent/MainContent';
-import Navbar from './navbar/NavbarLoggedIn';
-import Sidebar from './sidebar/Sidebar';
-
-function TimesheetMain() {
-
-    return (
-        <div className="app-container">
-           <Navbar />
-        </div>
-    )
-
-
-}
-
-export default TimesheetMain;
