+++ src/timetracker/assets/ClockIcon.jsx
