+++ src/timetracker/assets/HamburgerIcon.jsx
