+++ src/timetracker/assets/ProjectsIcon.jsx
