+++ src/timetracker/assets/logo.png
