+++ src/timetracker/components/navbar/NavbarLoggedIn.jsx
@@ -1,5 +1,3 @@
-import "./Navbar.css";
-
 import React from 'react';
 
 function Navbar() {
