+++ src/timetracker/pages/create-account-page/CreateAccountPage.css
@@ -5,17 +5,6 @@
   font-family: "Inter";
 }
 
-.create-account {
-  background-image: linear-gradient(
-      to bottom,
-      rgba(48, 188, 237, 100%),
-      rgba(63, 141, 169, 61.33%),
-      rgba(0, 102, 255, 25%)
-    ),
-    url(../img/background.png);
-  background-size: cover;
-}
-
 .create-account--greeting {
   text-shadow: 0 4px 4px rgba(0, 0, 0, 0.25);
 }
