+++ src/timetracker/pages/time-sheet/TimesheetMain.css
@@ -1,6 +1,4 @@
 .app-container {
-    background-image: url('/Public/Perlin1.png');
-    background-color: #30BCED;
     color: #fffaff;
     min-height: 100vh;
     padding: 3px;
@@ -8,7 +6,5 @@
     
     background-size: cover;
     background-repeat: no-repeat;
-
-    
 }
