+++ src/timetracker/pages/time-sheet/TimesheetMain.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import './TimesheetMain.css';
+import MainContent from '../MainContent/MainContent';
+
+function TimesheetMain() {
+    return (
+        <div className="">
+           <MainContent />
+        </div>
+    )
+}
+
+export default TimesheetMain;
